pyproject.toml
[tool.poetry]
name = "logchat"
version = "1.0.0"
description = ""
package-mode = false
authors = ["Hannes Ehringfeld <Hannes.Ehringfeld@ruhr-uni-bochum.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.8"
langchain = {extras = ["openai"], version = "^0.3.18"}
python-dotenv = "^1.0.1"
langgraph = "^0.2.70"
langchain-postgres = "^0.0.13"
langgraph-checkpoint-postgres = "^2.0.14"
typer = "^0.15.1"
sqlalchemy = "^2.0.38"
psycopg2-binary = "^2.9.10"
langchain-google-genai = "^2.0.11"
grpcio = "1.67.1"

[tool.poetry.group.dev.dependencies]
commitizen = "^4.4.1"
flask = "^3.1.0"
requests = "^2.32.3"
alembic = "^1.14.1"
ipykernel = "^6.29.5"



[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "pep440"
version_provider = "poetry"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
